update vault scripts

+++ docker/addons/vault/vault_set_pki.sh
@@ -24,9 +24,7 @@ if [ -n "${MG_NGINX_SERVER_NAME:-}" ]; then
     server_name="$MG_NGINX_SERVER_NAME"
 fi
 
-vault() {
-    docker exec -it magistrala-vault vault "$@"
-}
+source vault_cmd.sh
 
 vaultEnablePKI() {
     vault secrets enable -namespace=${MG_VAULT_NAMESPACE} -address=${MG_VAULT_ADDR}  -path ${MG_VAULT_PKI_PATH} pki
@@ -105,24 +103,43 @@ vaultGenerateIntermediateCSR() {
 
 vaultSignIntermediateCSR() {
     echo "Sign intermediate CSR"
-    docker cp data/${MG_VAULT_PKI_INT_FILE_NAME}.csr magistrala-vault:/vault/${MG_VAULT_PKI_INT_FILE_NAME}.csr
-    vault write -namespace=${MG_VAULT_NAMESPACE} -address=${MG_VAULT_ADDR} -format=json  ${MG_VAULT_PKI_PATH}/root/sign-intermediate \
-        csr=@/vault/${MG_VAULT_PKI_INT_FILE_NAME}.csr  ttl="8760h" \
-        ou="\"$MG_VAULT_PKI_INT_CA_OU\""\
-        organization="\"$MG_VAULT_PKI_INT_CA_O\"" \
-        country="\"$MG_VAULT_PKI_INT_CA_C\"" \
-        locality="\"$MG_VAULT_PKI_INT_CA_L\"" \
-        province="\"$MG_VAULT_PKI_INT_CA_ST\"" \
-        street_address="\"$MG_VAULT_PKI_INT_CA_ADDR\"" \
-        postal_code="\"$MG_VAULT_PKI_INT_CA_PO\"" \
-        | tee >(jq -r .data.certificate >data/${MG_VAULT_PKI_INT_FILE_NAME}.crt) \
-              >(jq -r .data.issuing_ca >data/${MG_VAULT_PKI_INT_FILE_NAME}_issuing_ca.crt)
+    if is_container_running "magistrala-vault"; then
+        docker cp data/${MG_VAULT_PKI_INT_FILE_NAME}.csr magistrala-vault:/vault/${MG_VAULT_PKI_INT_FILE_NAME}.csr
+        vault write -namespace=${MG_VAULT_NAMESPACE} -address=${MG_VAULT_ADDR} -format=json  ${MG_VAULT_PKI_PATH}/root/sign-intermediate \
+            csr=@/vault/${MG_VAULT_PKI_INT_FILE_NAME}.csr  ttl="8760h" \
+            ou="\"$MG_VAULT_PKI_INT_CA_OU\""\
+            organization="\"$MG_VAULT_PKI_INT_CA_O\"" \
+            country="\"$MG_VAULT_PKI_INT_CA_C\"" \
+            locality="\"$MG_VAULT_PKI_INT_CA_L\"" \
+            province="\"$MG_VAULT_PKI_INT_CA_ST\"" \
+            street_address="\"$MG_VAULT_PKI_INT_CA_ADDR\"" \
+            postal_code="\"$MG_VAULT_PKI_INT_CA_PO\"" \
+            | tee >(jq -r .data.certificate >data/${MG_VAULT_PKI_INT_FILE_NAME}.crt) \
+                >(jq -r .data.issuing_ca >data/${MG_VAULT_PKI_INT_FILE_NAME}_issuing_ca.crt)
+    else
+        vault write -namespace=${MG_VAULT_NAMESPACE} -address=${MG_VAULT_ADDR} -format=json  ${MG_VAULT_PKI_PATH}/root/sign-intermediate \
+            csr=@data/${MG_VAULT_PKI_INT_FILE_NAME}.csr  ttl="8760h" \
+            ou="\"$MG_VAULT_PKI_INT_CA_OU\""\
+            organization="\"$MG_VAULT_PKI_INT_CA_O\"" \
+            country="\"$MG_VAULT_PKI_INT_CA_C\"" \
+            locality="\"$MG_VAULT_PKI_INT_CA_L\"" \
+            province="\"$MG_VAULT_PKI_INT_CA_ST\"" \
+            street_address="\"$MG_VAULT_PKI_INT_CA_ADDR\"" \
+            postal_code="\"$MG_VAULT_PKI_INT_CA_PO\"" \
+            | tee >(jq -r .data.certificate >data/${MG_VAULT_PKI_INT_FILE_NAME}.crt) \
+                >(jq -r .data.issuing_ca >data/${MG_VAULT_PKI_INT_FILE_NAME}_issuing_ca.crt)
+    fi
+
 }
 
 vaultInjectIntermediateCertificate() {
     echo "Inject Intermediate Certificate"
-    docker cp data/${MG_VAULT_PKI_INT_FILE_NAME}.crt magistrala-vault:/vault/${MG_VAULT_PKI_INT_FILE_NAME}.crt
-    vault write -namespace=${MG_VAULT_NAMESPACE} -address=${MG_VAULT_ADDR} ${MG_VAULT_PKI_INT_PATH}/intermediate/set-signed certificate=@/vault/${MG_VAULT_PKI_INT_FILE_NAME}.crt
+    if is_container_running "magistrala-vault"; then
+        docker cp data/${MG_VAULT_PKI_INT_FILE_NAME}.crt magistrala-vault:/vault/${MG_VAULT_PKI_INT_FILE_NAME}.crt
+        vault write -namespace=${MG_VAULT_NAMESPACE} -address=${MG_VAULT_ADDR} ${MG_VAULT_PKI_INT_PATH}/intermediate/set-signed certificate=@/vault/${MG_VAULT_PKI_INT_FILE_NAME}.crt
+    else
+        vault write -namespace=${MG_VAULT_NAMESPACE} -address=${MG_VAULT_ADDR} ${MG_VAULT_PKI_INT_PATH}/intermediate/set-signed certificate=@data/${MG_VAULT_PKI_INT_FILE_NAME}.crt
+    fi
 }
 
 vaultGenerateIntermediateCertificateBundle() {
@@ -141,7 +158,7 @@ vaultSetupIntermediateIssuingURLs() {
 }
 
 vaultSetupServerCertsRole() {
-    if [ "$SKIP_SERVER_CERT" == "skip_server_cert" ]; then
+    if [ "$SKIP_SERVER_CERT" == "--skip-server-cert" ]; then
         echo "Skipping server certificate role"
     else
         echo "Setup Server certificate role"
@@ -152,7 +169,7 @@ vaultSetupServerCertsRole() {
 }
 
 vaultGenerateServerCertificate() {
-    if [ "$SKIP_SERVER_CERT" == "skip_server_cert" ]; then
+    if [ "$SKIP_SERVER_CERT" == "--skip-server-cert" ]; then
         echo "Skipping generate server certificate"
     else
         echo "Generate server certificate"
@@ -173,7 +190,9 @@ vaultSetupThingCertsRole() {
 }
 
 vaultCleanupFiles() {
-    docker exec magistrala-vault sh -c 'rm -rf /vault/*.{crt,csr}'
+    if is_container_running "magistrala-vault"; then
+        docker exec magistrala-vault sh -c 'rm -rf /vault/*.{crt,csr}'
+    fi
 }
 
 if ! command -v jq &> /dev/null
